low prog

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,35 +1,25 @@
-/*
- * File: 0-linear.c
- 
- */
-
 #include "search_algos.h"
 
 /**
-  * linear_search - Searches for a value in an array
-  *                 of integers using linear search.
-  * @array: A pointer to the first element of the array to search.
-  * @size: The number of elements in the array.
-  * @value: The value to search for.
+  * linear_search - linear search
+  * @array: pointer to first element in search array
+  * @size: size of array
+  * @value: value to find
   *
-  * Return: If the value is not present or the array is NULL, -1.
-  *         Otherwise, the first index where the value is located.
-  *
-  * Description: Prints a value every time it is compared in the array.
+  * Return: index where found or -1
   */
+
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	unsigned int i;
 
 	if (array == NULL)
 		return (-1);
-
 	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
+		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
-
 	return (-1);
 }
